(PC-2555): BSR -> re-organized profile folder
