+++ src/components/pages/profile/Profile.jsx
@@ -3,8 +3,8 @@ import React, { PureComponent } from 'react'
 import { Route, Switch } from 'react-router-dom'
 
 import { config } from './config'
-import ProfileMainView from './ProfileMainView'
-import ProfileUpdateSuccess from './ProfileUpdateSuccess'
+import ProfileMainView from './ProfileMainView/ProfileMainView'
+import ProfileUpdateSuccess from './ProfileUpdateSuccess/ProfileUpdateSuccess'
 import NotMatch from '../not-match/NotMatch'
 import LoaderContainer from '../../layout/Loader/LoaderContainer'
