+++ src/components/pages/profile/forms/fields/UserPasswordField.jsx
@@ -5,7 +5,7 @@ import FormError from '../../../../forms/FormError'
 import isEmpty from '../../../../../utils/strings/isEmpty'
 import PasswordField from '../../../../forms/inputs/PasswordField'
 import validateMatchingFields from '../../../../forms/validators/validateMatchingFields'
-import withProfileForm from '../withProfileForm'
+import withProfileForm from '../hocs/withProfileForm'
 
 const ERROR_OLD_PASSWORD = 'L’ancien mot de passe est manquant.'
